Refactor CommentService again into an interface with an implementing class DatastoreCommentService
